requirements.txt
sentence-transformers
uvicorn
transformers
ctransformers
langchain
langchain-community
python-box
streamlit
python-dotenv